Split out analyzer spelling golden tests
